Add AvroProvider and HiveProvider to unshim list
