+++ sql-plugin/src/main/scala/org/apache/spark/sql/hive/rapids/GpuHiveOverrides.scala
@@ -20,15 +20,6 @@ import com.nvidia.spark.rapids.{ExprRule, ShimLoader}
 
 import org.apache.spark.sql.catalyst.expressions.Expression
 
-/**
- * The subclass of HiveProvider imports spark-hive classes. This file should not imports
- * spark-hive because `class not found` exception may throw if spark-hive does not exist at
- * runtime. Details see: https://github.com/NVIDIA/spark-rapids/issues/5648
- */
-trait HiveProvider {
-  def getExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]]
-}
-
 object GpuHiveOverrides {
   def isSparkHiveAvailable: Boolean = {
     try {
